# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SkillForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Search users by skills or name
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 12
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user profile by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - skills
        - createdAt
    UserProfile:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
      required:
        - id
        - email
        - name
        - skills
        - createdAt
        - projects
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - users
        - total
        - page
        - limit
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        githubUrl:
          type: string
        skills:
          type: array
          items:
            type: string
    Project:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
        teamSize:
          type: number
        status:
          type: string
          enum: [looking_for_team, active, completed]
        githubUrl:
          type: string
          nullable: true
        ownerId:
          type: number
        ownerName:
          type: string
        ownerAvatarUrl:
          type: string
          nullable: true
        memberCount:
          type: number
        isBookmarked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - techStack
        - teamSize
        - status
        - ownerId
        - ownerName
        - memberCount
        - isBookmarked
        - createdAt
...